refactor: handler updates
- refactor handlers into auth and contacts subdirectories
- delete PoC handlers
