+++ rocket-mail/src/handlers/contacts-handlers/DeleteHandler.ts
